% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CFRET}
\alias{CFRET}
\title{Residual for observed mortality at time T}
\usage{
CFRET(T, km, Lm)
}
\arguments{
\item{T}{A 2xN matrix: the first column being observation times; the second column being fractional mortality at each time in the (control) cohort}

\item{km}{A proposed Weibull shape parameger}

\item{Lm}{A proposed Weibull scale parameter}
}
\value{
The error as a difference
}
\description{
Errors in cumulative mortality
}
\details{
This function is a utility for fitting a Weibull to the observed mortality in a control arm.
}
\author{
Pete Dodd
}
