% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CIfactor}
\alias{CIfactor}
\title{CIfactor}
\usage{
CIfactor(N, frac.deaths.control)
}
\arguments{
\item{N}{the number of deaths observed}

\item{frac.deaths.control}{the fraction of deaths that are expected in the control arm (returned by ITBstats)}
}
\value{
A factor for computation of approximate CIs. High = mid * factor; Low = mid / factor.
}
\description{
For computing a factor for approximate IRR CIs
}
\details{
For computing a factor for approximate IRR CIs.
}
\author{
Pete Dodd
}
