% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMortParz}
\alias{getMortParz}
\title{Get the best-fit parameters for mortality}
\usage{
getMortParz(M)
}
\arguments{
\item{M}{A 2xN matrix: the first column being observation times; the second column being fractional mortality at each time in the (control) cohort}
}
\value{
A list with a logical flag to indicate convergence, and the best-fit Weibull shape and scale parameters.
}
\description{
For fitting to mortality data.
}
\details{
This takes data from a control arm on cumulative mortality and fits a Weibull distribution to it so as to minimize the sum-of-squares error.
}
\author{
Pete Dodd
}
