% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTxParz}
\alias{getTxParz}
\title{Get the best-fit parameters for treatment}
\usage{
getTxParz(M, km, lm)
}
\arguments{
\item{M}{A 2xN matrix: the first column being observation times; the second column being cumulative fraction treated at each time in the cohort of those who are ultimately treated}

\item{km}{The best-fit Weibull shape parameter for mortality}

\item{lm}{The best-fit Weibull scale parameter for mortality}
}
\value{
A list with a logical flag to indicate convergence, and the best-fit Weibull shape and scale parameters.
}
\description{
For fitting to treatment data.
}
\details{
This takes data on cumulative percentiles of cumulative treatment among those treated, and fits a Weibull distribution to it so as to minimize the sum-of-squares error, taking into account the competing hazard of death.
}
\author{
Pete Dodd
}
