% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ITBstats}
\alias{ITBstats}
\title{ITBstats}
\usage{
ITBstats(
  N = 10000,
  Tstop = 90,
  Tlandmark = 1,
  Texc = 0,
  rtt.exposure,
  rtt.death,
  rtt.ltfu,
  returnraw = FALSE,
  ...
)
}
\arguments{
\item{N}{The size of the cohort for simulation (unrelated to the experimental study cohort size)}

\item{Tstop}{The time at which observation stops}

\item{Tlandmark}{The time used for a landmark analysis}

\item{Texc}{The time used for exclude early events & time analysis}

\item{rtt.exposure}{random time-to-exposure generator}

\item{rtt.death}{random time-to-death generator}

\item{rtt.ltfu}{random time-to-ltfu generator}

\item{returnraw}{Logical (default FALSE) controlling return type. If true, a list comprising the raw cohort and the raw landmark cohort will be returned. Otherwise, a list of summary statistics will be returned}
}
\value{
Depends on [returnraw].

Note that if returnraw==FALSE, the following labels are used for analysis variants:
a) person time from 0 (standard); b) excluding early events & don't reset clock; c) landmark analysis: drop those dead/ltfu before landmark AND exposed after landmark -> control; d) excluding early events & reset clock
}
\description{
Function to simulate a cohort
}
\details{
This function will simulate a cohort with relevant times-to-event under a null assumption for the treatment effect. Typical use returns summary statistics.
}
\author{
Pete Dodd
}
