% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDistPlot}
\alias{makeDistPlot}
\title{makeDistPlot}
\usage{
makeDistPlot(input)
}
\arguments{
\item{inputs}{a list of parameters containing the Weibull parameters for exposure (k.e, L.e), death (k.d, L.d), and ltfu (k.l, L.l)}

\item{k.e}{(tmp)}

\item{k.d}{(tmp)}

\item{k.l}{(tmp)}

\item{N}{The size of the cohort for simulation (unrelated to the experimental study cohort size)}

\item{rtt.exposure}{random time-to-exposure generator}

\item{rtt.death}{random time-to-death generator}

\item{rtt.ltfu}{random time-to-ltfu generator}

\item{returnraw}{Logical (default FALSE) controlling return type. If true, the raw cohort will be returned. Otherwise, a list of summary statistics will be returned}
}
\value{
A ggplot2 graph
}
\description{
Function to visualize input distributions
}
\details{
This plots the exposure (treatment), mortality, and LTFU time-to-event distributions
}
\author{
Pete Dodd
}
