% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeEffectBySamplePlot}
\alias{makeEffectBySamplePlot}
\title{make Effect by Sample size Plot}
\usage{
makeEffectBySamplePlot(ANS)
}
\arguments{
\item{ANS}{A list of summary statistics returned by [ITBstats]}
}
\value{
A ggplot2 graph
}
\description{
Make plot of effect and CIs by sample size
}
\details{
This is the effect by various analyses, with approximate confidence intervals, plotted against sample size.
}
\author{
Pete Dodd
}
