% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeFitCheckPlot}
\alias{makeFitCheckPlot}
\title{makeFitCheckPlot}
\usage{
makeFitCheckPlot(input, FT, FV)
}
\arguments{
\item{input}{a list of parameters: the mortality times (mt), fractions (mf), and treatment times (tt), fractions (tf).}

\item{FT}{a list of parameters including the maximum time (T.max), denominator type (denominator)}

\item{FV}{a list of parameters including Weibull treatment distribution parameters (k.e, L.e), Weibull mortality distribution parameters (k.d, L.d)}
}
\value{
A plot of data and cumulative mortality and exposure.
}
\description{
Plot survival and treatment distributions.
}
\details{
This is a utility to check fitted survival distributions against data for plausibility.
}
\author{
Pete Dodd
}
