% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeResultList}
\alias{makeResultList}
\title{Make Results List}
\usage{
makeResultList(
  mortality.times,
  mortality.fracs,
  treatment.times,
  treatment.fracs,
  N,
  simulation.cohort.size = 1e+05,
  Tmax = 30,
  Tlandmark = 1,
  Tearly = 1,
  plotfit = FALSE,
  denominator = "cohort",
  ...
)
}
\arguments{
\item{mortality.times}{a vector of mortality data times}

\item{mortality.fracs}{a vector of mortality fractions among all patients corresponding to times}

\item{treatment.times}{a vector of treatment data times}

\item{treatment.fracs}{a vector of corresponding treatment fractions among all patients}

\item{N}{the number of deaths observed (for CIs)}

\item{simulation.cohort.size}{the size of the cohort used in the simulation}

\item{Tmax}{the maximum time horizon}

\item{Tlandmark}{the time used in landmark analysis}

\item{Tearly}{the time used in the drop-early-events & time analysis}

\item{plotfit}{Plot the Weibull fits?}

\item{denominator}{Whether fractions have the whole cohort as denominator (default: 'cohort') or only those ultimately 'exposed'}

\item{...}{any extras}

\item{frac.deaths.control}{the fraction of deaths that are expected in the control arm (returned by ITBstats)}
}
\value{
A list of results from ITBstats together with the number of deaths
}
\description{
A shortcut function to generate all results
}
\details{
This is a shortcut function to fit to mortality and treatment data and return results. This uses a cohort of 10,000 by default.
}
\author{
Pete Dodd
}
