% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeTMplot}
\alias{makeTMplot}
\title{Plot survival and treatment distributions.}
\usage{
makeTMplot(input)
}
\arguments{
\item{input}{a list of parameters including Weibull treatment distribution parameters (k.e, L.e), Weibull mortality distribution parameters (k.d, L.d), and the maximum time (T.max).}
}
\value{
A plot with survival and treatment times
}
\description{
Plot survival and treatment distributions.
}
\details{
The plot returned is the survival (from death) function with a scaled time-to-treatment distribution superimposed. Note: the time-to-treatment distribution will differ from the time-to-treatment distribution in the data due to competition with mortality. The competition with mortality is accounted for in fitting and simulation, but not in this plot.
}
\author{
Pete Dodd
}
