% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{morterr}
\alias{morterr}
\title{SSE for fitting a Weibull to observed mortality}
\usage{
morterr(M, x)
}
\arguments{
\item{M}{A 2xN matrix: the first column being observation times; the second column being fractional mortality at each time in the (control) cohort}

\item{x}{A 2-vector equal to (log(Weibull shape), log(Weibull scale))}
}
\value{
The SSE for the Weibull and these data
}
\description{
SSE mortality errors
}
\details{
This is a utility function for fitting a Weibull to observed mortality in a control arm..
}
\author{
Pete Dodd
}
