% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Yfit}
\alias{Yfit}
\title{Fit Weilbull}
\usage{
Yfit(t, F)
}
\arguments{
\item{t}{vector of time points}

\item{F}{vector of fractions with event}
}
\value{
vector (k,L) of Weibull shape/scale parameters
}
\description{
Fit Weibull to times and fractions
}
\details{
This uses an OLS fit to transformed Weibull parameters to fit. See Weibull distribution Wikipedia page for explanation.
}
\examples{
mortality.times <- c(2,8,30)
mortality.fracs <- c(0.07, 0.14, 0.14)
mortality.parms <- Yfit(mortality.times, mortality.fracs)

}
\author{
Pete Dodd
}
