% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCItable}
\alias{makeCItable}
\title{Make a table of results}
\usage{
makeCItable(list.of.results)
}
\arguments{
\item{list.of.results}{a named list of results from makeResultList, ie from ITBstats with N=#deaths appended}
}
\value{
a data frame of results for all elements in the list
}
\description{
For generating a results table from a list of individual study results
}
\details{
Generates a results table including CIs for analyses a)-d) of ITB stats as well as the fraction of deaths expected to occur in the control arm.
}
\author{
Pete Dodd
}
