% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeFitCheckPlot}
\alias{makeFitCheckPlot}
\title{makeFitCheckPlot}
\usage{
makeFitCheckPlot(input)
}
\arguments{
\item{input}{a list of parameters including Weibull treatment distribution parameters (k.e, L.e), Weibull mortality distribution parameters (k.d, L.d), and the maximum time (T.max), denominator type, and the mortality and treatment data.}
}
\value{
A plot
}
\description{
Plot survival and treatment distributions.
}
\details{
This is a utility to check fitted survival distributions against data for plausibility.
}
\author{
Pete Dodd
}
